DamageText 크리티컬 떴을때 색 변경

+++ Assets/Scripts/Operator/CriticalOperator.cs
@@ -9,7 +9,7 @@ public static class CriticalOperator
 
     public static int GetCriticalDamageIntValue(float damage, float criticalValue)
     {
-        return (int)damage * (int)criticalValue;
+        return (int)(damage * criticalValue);
     }
 
     public static float GetCriticalDamageFloatValue(float damage, float criticalValue)
